requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
requests
python-dotenv
rich
keyboard
pyperclip
